// frontend/src/api/generated/schemas/timeMode.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.

/**
 * Time mode for data analysis.
 */
export type TimeMode = (typeof TimeMode)[keyof typeof TimeMode];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TimeMode = {
  snapshot: "snapshot",
  period: "period",
} as const;